+++ internal/adventuria/actions/custom/drop/action.go
@@ -103,8 +103,8 @@ func (d *Drop) CanDo(ctx context.Context, events *model.Events, player *model.Pl
 }
 
 type Request struct {
-	Comment string `json:"comment"`
-	Score   int    `json:"score"`
+	Comment string  `json:"comment"`
+	Score   float64 `json:"score"`
 }
 
 func (d *Drop) Do(ctx context.Context, events *model.Events, player *model.Player, actionReq model.ActionRequest) (any, error) {
